fix: update mean_correct_kernel to use long long for dimensions and indices
- Changed the data types for `n_obs` and `k` in `mean_correct_kernel` from `int` to `long long` to accommodate larger matrix sizes.
- Updated index calculations within the kernel to use `long long` for `idx` and `col`, ensuring compatibility with the new data types.
- Adjusted the `mean_correct_gpu` function to reflect these changes, including type conversions for `n_obs` and `k` to `i64` and validation for CUDA grid dimensions to prevent overflow errors.
- Enhanced error handling for block calculations to ensure they remain within the limits of CUDA's grid dimensions.

+++ scx-gpu/kernels/spmm_mean_correct.cu
@@ -10,12 +10,12 @@
 extern "C" __global__ void mean_correct_kernel(
     float* __restrict__ Y,         // [n_obs × k], row-major
     const float* __restrict__ mc,  // [k], mean correction vector
-    int n_obs, int k
+    long long n_obs, long long k
 ) {
-    int idx = blockIdx.x * blockDim.x + threadIdx.x;
-    int total = n_obs * k;
+    long long idx = (long long)blockIdx.x * blockDim.x + threadIdx.x;
+    long long total = n_obs * k;
     if (idx >= total) return;
 
-    int col = idx % k;
+    long long col = idx % k;
     Y[idx] -= mc[col];
 }
